Fix import issues from `packages/frontend/src/utilities.ts`
- Moved the types into a different file
