+++ packages/functions/src/utilities/fullTestUtilities.ts
@@ -1,4 +1,3 @@
-///  <reference path="../../../frontend/src/utilities.ts" />
 ///  <reference path="../../../frontend/src/utilities/LRUtilities.ts" />
 import { LRQuestion } from '../../../frontend/src/utilities/LRUtilities';
